requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
requests>=2.26.0
openai>=1.0.0
openpipe>=0.1.0
Pillow>=9.0.0
pytesseract>=0.3.8
python-magic>=0.4.24
backoff>=2.2.1
httpx>=0.25.0
